Rename to slot-block
